Fix server issues
