+++ src/main/java/com/bobmowzie/mowziesmobs/server/capability/AbilityCapability.java
@@ -204,123 +204,4 @@ public class AbilityCapability {
             return cap == ABILITY_CAPABILITY ? instance.cast() : LazyOptional.empty();
         }
     }
-
-    public static final class AbilityEventHandler {
-        @SubscribeEvent
-        public void onPlayerInteract(PlayerInteractEvent.RightClickEmpty event) {
-            PlayerEntity player = event.getPlayer();
-            IAbilityCapability abilityCapability = AbilityHandler.INSTANCE.getAbilityCapability(player);
-            if (abilityCapability != null) {
-                for (Ability ability : abilityCapability.getAbilities()) {
-                    ability.onRightClickEmpty(event);
-                }
-            }
-        }
-
-        @SubscribeEvent
-        public void onPlayerRightClickBlock(PlayerInteractEvent.RightClickBlock event) {
-            PlayerEntity player = event.getPlayer();
-            IAbilityCapability abilityCapability = AbilityHandler.INSTANCE.getAbilityCapability(player);
-            if (abilityCapability != null) {
-                for (Ability ability : abilityCapability.getAbilities()) {
-                    ability.onRightClickBlock(event);
-                }
-            }
-        }
-
-        @SubscribeEvent
-        public void onPlayerRightClickItem(PlayerInteractEvent.RightClickItem event) {
-            PlayerEntity player = event.getPlayer();
-            IAbilityCapability abilityCapability = AbilityHandler.INSTANCE.getAbilityCapability(player);
-            if (abilityCapability != null) {
-                for (Ability ability : abilityCapability.getAbilities()) {
-                    ability.onRightClickWithItem(event);
-                }
-            }
-        }
-
-        @SubscribeEvent
-        public void onPlayerRightClickEntity(PlayerInteractEvent.EntityInteract event) {
-            PlayerEntity player = event.getPlayer();
-            IAbilityCapability abilityCapability = AbilityHandler.INSTANCE.getAbilityCapability(player);
-            if (abilityCapability != null) {
-                for (Ability ability : abilityCapability.getAbilities()) {
-                    ability.onRightClickEntity(event);
-                }
-            }
-        }
-
-        @SubscribeEvent
-        public void onPlayerLeftClickEmpty(PlayerInteractEvent.LeftClickEmpty event) {
-            PlayerEntity player = event.getPlayer();
-            IAbilityCapability abilityCapability = AbilityHandler.INSTANCE.getAbilityCapability(player);
-            if (abilityCapability != null) {
-                for (Ability ability : abilityCapability.getAbilities()) {
-                    ability.onLeftClickEmpty(event);
-                }
-            }
-        }
-
-        @SubscribeEvent
-        public void onPlayerLeftClickBlock(PlayerInteractEvent.LeftClickBlock event) {
-            PlayerEntity player = event.getPlayer();
-            IAbilityCapability abilityCapability = AbilityHandler.INSTANCE.getAbilityCapability(player);
-            if (abilityCapability != null) {
-                for (Ability ability : abilityCapability.getAbilities()) {
-                    ability.onLeftClickBlock(event);
-                }
-            }
-        }
-
-        @SubscribeEvent
-        public void onLeftClickEntity(AttackEntityEvent event) {
-            PlayerEntity player = event.getPlayer();
-            IAbilityCapability abilityCapability = AbilityHandler.INSTANCE.getAbilityCapability(player);
-            if (abilityCapability != null) {
-                for (Ability ability : abilityCapability.getAbilities()) {
-                    ability.onLeftClickEntity(event);
-                }
-            }
-        }
-
-        @SubscribeEvent
-        public void onTakeDamage(LivingHurtEvent event) {
-            LivingEntity player = event.getEntityLiving();
-            IAbilityCapability abilityCapability = AbilityHandler.INSTANCE.getAbilityCapability(player);
-            if (abilityCapability != null) {
-                for (Ability ability : abilityCapability.getAbilities()) {
-                    ability.onTakeDamage(event);
-                }
-            }
-        }
-
-        @SubscribeEvent
-        public void onJump(LivingEvent.LivingJumpEvent event) {
-            LivingEntity player = event.getEntityLiving();
-            IAbilityCapability abilityCapability = AbilityHandler.INSTANCE.getAbilityCapability(player);
-            if (abilityCapability != null) {
-                for (Ability ability : abilityCapability.getAbilities()) {
-                    ability.onJump(event);
-                }
-            }
-        }
-    }
-
-    @OnlyIn(Dist.CLIENT)
-    public enum AbilityClientEventHandler {
-        INSTANCE;
-
-        @SubscribeEvent
-        public void onRenderTick(TickEvent.RenderTickEvent event) {
-            PlayerEntity player = Minecraft.getInstance().player;
-            if (player != null) {
-                IAbilityCapability abilityCapability = AbilityHandler.INSTANCE.getAbilityCapability(player);
-                if (abilityCapability != null) {
-                    for (Ability ability : abilityCapability.getAbilities()) {
-                        ability.onRenderTick(event);
-                    }
-                }
-            }
-        }
-    }
 }
